Refactoring util methods.
