+++ tests/utils.js
@@ -4,7 +4,8 @@
  */
 
 import ViewElement from 'ckeditor5-engine/src/view/element';
-import { toImageWidget, isImageWidget } from 'ckeditor5-image/src/utils';
+import ModelElement from 'ckeditor5-engine/src/model/element';
+import { toImageWidget, isImageWidget, isImage } from 'ckeditor5-image/src/utils';
 import { isWidget } from 'ckeditor5-image/src/widget/utils';
 
 describe( 'image widget utils', () => {
@@ -28,6 +29,25 @@ describe( 'image widget utils', () => {
 
 		it( 'should return false for non-widgetized elements', () => {
 			expect( isImageWidget( new ViewElement( 'p' ) ) ).to.be.false;
+		} );
+	} );
+
+	describe( 'isImage', () => {
+		it( 'should return true for image element', () => {
+			const image = new ModelElement( 'image' );
+
+			expect( isImage( image ) ).to.be.true;
+		} );
+
+		it( 'should return true false for different elements', () => {
+			const image = new ModelElement( 'foo' );
+
+			expect( isImage( image ) ).to.be.false;
+		} );
+
+		it( 'should return true false for null and undefined', () => {
+			expect( isImage( null ) ).to.be.false;
+			expect( isImage( undefined ) ).to.be.false;
 		} );
 	} );
 } );
